Remove IReadOnlyList implementation from SdbDataArray
There's no need to abstract it, and this will remove any virtual calls which can add up for performance reasons.
