+++ src/DocumentFormat.OpenXml/DocumentFormat.OpenXml.csproj
@@ -36,7 +36,6 @@
       </ItemGroup>
 
       <ItemGroup>
-        <Compile Include="System\Collections\Generic\IReadOnlyList.cs" />
         <Compile Include="System\Collections\Generic\IReadOnlyCollection.cs" />
         <Compile Include="System\Collections\Generic\IReadOnlyDictionary.cs" />
         <Compile Include="System\Reflection\TypeExtensions.cs" />
